trader_repo/pyproject.toml

[project]
name = "trader"
version = "0.1.0"
description = "Research, backtesting, walk-forward, and live paper trading for US equities."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{name="Your Name"}]
dependencies = [
    "pandas==2.2.2",
    "numpy==1.26.4",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "typer==0.12.3",
    "yfinance==0.2.41",
    "pyyaml==6.0.2",
    "loguru==0.7.2",
    "matplotlib==3.8.4",
    "plotly==5.23.0",
    "optuna==3.6.1",
    "pandas-market-calendars==4.3.2",
    "tqdm==4.66.5",
    "alpaca-py==0.20.2",
    "reportlab==4.2.3"
]

[project.optional-dependencies]
dev = [
    "ruff==0.5.7",
    "black==24.8.0",
    "mypy==1.10.0",
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "types-PyYAML==6.0.12.20240808",
    "types-python-dateutil==2.9.0.20240316"
]

[project.scripts]
trader = "trader.__main__:app"

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
line-length = 100
lint.select = ["E","F","I","B","UP"]
lint.ignore = ["E501"]
target-version = "py311"
src = ["trader", "tests", "scripts"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.uv]
scripts.setup = "pre-commit install || true"
scripts.lint = "ruff check . && black --check . && mypy ."
scripts.test = "pytest -q"
